Further Delays
Due to a paid project I was hired for (in addition to a 2-day sickness
last week), the project release day in postponed for the first days of
November. This time, I worked on the tutorials, doing most of front and
back end work, but the full tutorial text is still to be created (and
will be done last anyway).